Move types to where they're derived from.
